pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = ""
authors = [
    {name = "Dmitriy",email = "Krivopusk.mitya@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "pydantic (>=2.11.7,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "qrcode[pil] (>=8.2,<9.0)",
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "asyncpg>=0.29.0",
    "redis>=6.4.0",
    "alembic>=1.16.4",
    "python-dateutil>=2.9.0.post0",
    "web3>=7.13.0",
    "celery>=5.5.3",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
